# api/openapi.yaml
openapi: 3.0.3
info:
  title: Heromail API
  version: 0.1.0
  description: |
    Corporate mail service.

    This document is the source of truth for the HTTP contract. The Go server
    interface is generated from it, so an endpoint declared here that nobody
    implements fails to compile — the spec cannot quietly drift from the code.

    Two independent bearer credentials exist, and they are not interchangeable:

    * **session** — a signed-in person, issued by `/auth/login`. Guards `/auth/me`,
      `/auth/logout` and everything under `/workspaces`.
    * **apiKey** — a machine acting for one workspace, `hm_live_…`. Guards `/v1/*`.
servers:
  - url: https://api-dev-heromail.rnm.dev
    description: dev
  - url: http://localhost:8060
    description: local
...
security: []
paths:
...
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: The user behind the current session
      security: [{ session: [] }]
      responses:
        '200':
          description: The signed-in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke the current session
      security: [{ session: [] }]
      responses:
        '204':
          description: Session revoked
        '401': { $ref: '#/components/responses/Unauthorized' }
  /workspaces:
    get:
      tags: [workspaces]
      operationId: listWorkspaces
      summary: Workspaces the signed-in user belongs to
      security: [{ session: [] }]
      responses:
        '200':
          description: The user's workspaces, with their role in each
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      description: The creator becomes its owner.
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkspaceRequest' }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: The slug is taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /messages/search:
    get:
      tags: [emails]
      operationId: searchMessages
      summary: Search messages visible to the signed-in user
      description: Searches stored outbound messages across all workspaces the user belongs to.
      security: [{ session: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Matching messages, newest first
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Email' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      description: Session token from `/auth/register` or `/auth/login`.
    apiKey:
      type: http
      scheme: bearer
      description: Workspace API key, `hm_live_…`, minted by the issue-api-key command.
  responses:
    BadRequest:
      description: The request is malformed or fails validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, malformed, expired or revoked credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        The single error envelope for the whole API. `/v1/*` used to return a
        bare `{"error":"text"}`, which the frontend client could not read.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable. Branch on this, not on the message.
              example: validation_failed
            message:
              type: string
              description: Human-readable, safe to show a user.
            details:
              description: Optional per-field or per-code extra context.
...
    Slug:
      type: string
      pattern: '^[a-z0-9][a-z0-9-]{1,61}[a-z0-9]$'
      description: Lowercase tenant handle used in URLs.
      example: acme
    Role:
      type: string
      enum: [owner, admin, member]
    User:
      type: object
      required: [id, email, name, email_verified, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        email_verified:
          type: boolean
          description: Derived from `email_verified_at`; the flag the UI wants.
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserEnvelope:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/User' }
...
    Workspace:
      type: object
      required: [id, slug, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        slug: { $ref: '#/components/schemas/Slug' }
        name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceMembership:
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          required: [role]
          properties:
            role: { $ref: '#/components/schemas/Role' }
    WorkspaceList:
      type: object
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items: { $ref: '#/components/schemas/WorkspaceMembership' }
    CreateWorkspaceRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { $ref: '#/components/schemas/Slug' }
...
    EmailStatus:
      type: string
      enum: [queued, processing, sent, failed, dead]
      description: |
        `sent` and `dead` are terminal. Delivery, bounces and complaints are
        events that happen afterwards, not statuses.
...
    Attachment:
      type: object
      required: [id, filename, content_type, size_bytes, created_at]
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
...
    Email:
      type: object
      required: [id, workspace_id, from, to, subject, status, attempts, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        from: { type: string, format: email }
        to:
          type: array
          items: { type: string, format: email }
        subject: { type: string }
        html: { type: string, nullable: true }
        text: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/EmailStatus' }
        provider_message_id: { type: string, nullable: true }
        idempotency_key: { type: string, nullable: true }
        attempts: { type: integer }
        last_error: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/Attachment' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }